Changed base58 prefixes of secret key to match vanitygen convention
To be compatible with vanitygen (and others),
the prefix of SECRET_KEY should be the prefix of PUBKEY_ADDRESS + 128.
This also has the benefit of making it unambiguous that an address
is intended for use with Ribbitcoin, if user encounters
base58 secret keys (such as when redeeming a paper wallet).